pyproject.toml
[tool.poetry]
name = "imagazine"
version = "0.1.0"
description = ""
authors = ["Farid Zainulin <10221278@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "4.2.8"
python-dotenv = "^1.0.0"
django-bootstrap-v5 = "^1.0.11"
pillow = "^10.2.0"
django-filter = "^23.5"
django-bootstrap-form = "^3.4"
django-extensions = "2.1.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
